fix: moved tests to correct file
